--- checkIn/schema_updates/2017-10-09.sql
@@ -1,2 +1,0 @@
-ALTER TABLE locations ADD secret BINARY(16);
-ALTER TABLE locations ADD salt BINARY(16);
