--- checkIn/schema_updates/2017-10-27.sql
@@ -1,1 +1,0 @@
-ALTER TABLE users ADD pin BIGINT;
