--- checkIn/schema_updates/2017-11-03.sql
@@ -1,3 +1,0 @@
-UPDATE users SET pin=NULL WHERE pin IS NOT NULL;
-ALTER TABLE users MODIFY pin BINARY(32);
-ALTER TABLE users ADD COLUMN pin_salt BINARY(16);
